posts_service/pyproject.toml
[tool.poetry]
name = "post-service"
version = "0.1.0"
description = ""
authors = ["josericardopenase <josericardopenase@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.21"
pylint = "^2.17.5"
peewee = "^3.16.3"
neovim = "^0.3.1"
jedi = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
